refactor(ios): unify game script profiling
